RC values as uint16

+++ inc/msp/FlightController.hpp
@@ -220,10 +220,10 @@ public:
 
     bool isArmed();
 
-    bool setRc(const uint roll, const uint pitch,
-               const uint yaw, const uint throttle,
-               const uint aux1 = 1000, const uint aux2 = 1000,
-               const uint aux3 = 1000, const uint aux4 = 1000);
+    bool setRc(const uint16_t roll, const uint16_t pitch,
+               const uint16_t yaw, const uint16_t throttle,
+               const uint16_t aux1 = 1000, const uint16_t aux2 = 1000,
+               const uint16_t aux3 = 1000, const uint16_t aux4 = 1000);
 
     bool setMotors(const std::array<uint16_t,msp::N_MOTOR> &motor_values);
